@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
 import 'quote.dart';
 
-void main() => runApp(MaterialApp(
-  home: QuoteList()
-));
+void main() => runApp(MaterialApp(home: QuoteList()));
 
 class QuoteList extends StatefulWidget {
+  const QuoteList({Key? key}) : super(key: key);
+
   @override
   _QuoteListState createState() => _QuoteListState();
 }
 
 class _QuoteListState extends State<QuoteList> {
-
   //  List<String> quotes = [
   //    'Be yourself; everyone else is already taken',
   //    'I have nothing to declare except my genius',
@@ -19,9 +18,15 @@ class _QuoteListState extends State<QuoteList> {
   //  ];
 
   List<Quote> quotes = [
-    Quote(author: 'Oscar Wilde', text: 'Be yourself; everyone else is already taken'),
-    Quote(author: 'Oscar Wilde', text: 'I have nothing to declare except my genius'),
-    Quote(author: 'Oscar Wilde', text: 'The truth is rarely pure and never simple')
+    Quote(
+        author: 'Oscar Wilde',
+        text: 'Be yourself; everyone else is already taken'),
+    Quote(
+        author: 'Oscar Wilde',
+        text: 'I have nothing to declare except my genius'),
+    Quote(
+        author: 'Oscar Wilde',
+        text: 'The truth is rarely pure and never simple')
   ];
 
   @override
@@ -34,11 +39,10 @@ class _QuoteListState extends State<QuoteList> {
         backgroundColor: Colors.redAccent,
       ),
       body: Column(
-        children: quotes.map((quote) => Text('${quote.text} - ${quote.author}')).toList(),
+        children: quotes
+            .map((quote) => Text('${quote.text} - ${quote.author}'))
+            .toList(),
       ),
     );
   }
 }
-
-
-
